+++ templates/database/seeders/ExportSeeder.php
@@ -16,7 +16,7 @@ class ExportSeeder extends Seeder
      */
     public function run(): void
     {
-        Menuitem::insertIfNotExists('imports', 'bi bi-file-earmark-arrow-down');
+        Menuitem::insertIfNotExists('imports', 'bi bi-file-earmark-arrow-down', 'Import', '/export-index');
         Module::insertIfNotExists('Exports');
     }
 }